feat: query metric
