+++ server/apm/apm-server/apm-engine/src/main/java/io/holoinsight/server/apm/engine/storage/MetricStorage.java
@@ -19,7 +19,7 @@ public interface MetricStorage {
   List<String> listMetrics();
 
   MetricValues queryMetric(String tenant, String metric, Duration duration,
-      Map<String, Object> conditions, List<String> groups) throws IOException;
+      Map<String, Object> conditions, List<String> groups) throws Exception;
 
   List<String> querySchema(String metric);
