+++ server/apm/apm-server/apm-service/src/main/java/io/holoinsight/server/apm/server/service/impl/MetricServiceImpl.java
@@ -30,7 +30,7 @@ public class MetricServiceImpl implements MetricService {
 
   @Override
   public MetricValues queryMetric(String tenant, String metric, Duration duration,
-      Map<String, Object> conditions) throws IOException {
+      Map<String, Object> conditions) throws Exception {
     return metricStorage.queryMetric(tenant, metric, duration, conditions, null);
   }
